moved fstree package after rebase
