+++ src/fstree/Node.py
